+++ sources/shared/shared/plugin/yPluginApi/StandardCapacities.cpp
@@ -24,7 +24,7 @@ namespace shared
          DECLARE_CAPACITY(CStandardCapacities::Dimmable, "dimmable", CStandardUnits::Percent(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Direction, "direction", CStandardUnits::Degrees(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Distance, "distance", CStandardUnits::Meter(), EKeywordDataType::kNumeric);
-         DECLARE_CAPACITY(CStandardCapacities::Duration, "duration", CStandardUnits::NoUnit(), EKeywordDataType::kNumeric);
+         DECLARE_CAPACITY(CStandardCapacities::Duration, "duration", CStandardUnits::Second(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::ElectricLoad, "electricLoad", CStandardUnits::AmperePerHOur(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Energy, "energy", CStandardUnits::WattPerHour(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Event, "event", CStandardUnits::NoUnit(), EKeywordDataType::kNoData);